Add variable batch size/cpu only tests
